Convert unit test project to XUnit
