+++ Tenduke.Client/AssemblyInfo.cs
@@ -1,0 +1,7 @@
+﻿using System;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+
+[assembly: ComVisible(false)]
+[assembly: CLSCompliant(true)]
+[assembly: InternalsVisibleTo("Tenduke.Client.Tests, PublicKey=0024000004800000940000000602000000240000525341310004000001000100011d23be5f30d0faddc638c4a5dbdc8f6b584ce6403f14604e3abaf8864dc10b179f14dbb1c0cf9f472b464a11bc4cb9bdae3a3276b00c17d249b7a7478a0fc8bbee45db6059e4438f67296e00c4734ba297ec49ed8fe4b71a226fb1220e5f2de202e2232c5827804cc1fef34aaf2a13b5b3ea6e579a0835b326115d6db36bde")]
